requirement.txt
google-generativeai
streamlit
python-dotenv
langchain_community 
pymysql 
chromadb